Put slick projection back.
The overhead of keeping separate database configuration for journal and
projection is not worth it.

Yes, technically these are and should be two separate databases. Not for
the small services I write, though - and even if they would be, which
makes sense, it would still be easier to use the same method of
cofiguration for both.

Maybe drop Slick from the journal as well? Well, the official
akka-persistence-jdbc is using it, so I guess not, unless I want to
replace the JDBC altogether.